Generate summary based on export
Replace code to create the human readable summary from push with the output of
the table function that uses the same report export as the JSON or YAML output.

Move function to create a human readable time duration into the `cf` package
since it is needed in two different places now.